==> 06.1 - Combine county population estimates.ipynb <==
import pandas as pd

pop_2000 = pd.read_csv("data/raw/co-est00int-tot.csv", encoding="latin-1")

pop_2000["FIPS"] = pop_2000[["STATE", "COUNTY"]].apply(lambda row: "%0.2d%0.3d" % (row["STATE"], row["COUNTY"]), axis=1)

pop_2000.set_index("FIPS", inplace=True)

pop_2010 = pd.read_csv("data/raw/co-est2017-alldata.csv", encoding="latin-1")

pop_2010["FIPS"] = pop_2010[["STATE", "COUNTY"]].apply(lambda row: "%0.2d%0.3d" % (row["STATE"], row["COUNTY"]), axis=1)

pop_2010.set_index("FIPS", inplace=True)

pop_2010.head()

df = pop_2000.join(pop_2010, how="inner", rsuffix="_2010")

columns = [
    "POPESTIMATE%d" % (year)
    for year in range(2004,2015)
]

df = df[df["SUMLEV"] == 50]

df = df[columns]

f = open("data/processed/county_intersection_list_2004_2014.txt","r")
fips = f.read().strip().split("\n")
f.close()

df = df.loc[fips]

df.to_csv("data/processed/county_population_2004_2014.csv", index_label="FIPS")
